Fix printf-format issues

+++ code/cgame/cg_main.c
@@ -282,7 +282,7 @@ Q_EXPORT intptr_t vmMain( intptr_t command, intptr_t arg0, intptr_t arg1, intptr
 		return 0;
 
 	default:
-		CG_Error( "vmMain: unknown command %i", command );
+		CG_Error( "vmMain: unknown command %i", (int)command );
 		break;
 	}
 	return -1;
