+++ code/game/g_session.c
@@ -68,7 +68,7 @@ void G_WriteClientSessionData( gclient_t *client ) {
 		client->sess.referee
 		);
 
-	var = va( "session%i", client - level.clients );
+	var = va( "session%i", (int)(client - level.clients) );
 
 	trap_Cvar_Set( var, s );
 }
@@ -94,7 +94,7 @@ void G_ReadSessionData( gclient_t *client ) {
 	int motdSeen;
 	int referee;
 
-	var = va( "session%i", client - level.clients );
+	var = va( "session%i", (int)(client - level.clients) );
 	trap_Cvar_VariableStringBuffer( var, s, sizeof(s) );
 
 	sscanf( s, "%i %i %i %i %i %i %i %i %i %i %i %i",
